requirements.txt
asgiref==3.8.1
attrs==23.2.0
automium==0.2.6
automium_web==0.1.1
blinker==1.7.0
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.4
djangorestframework==3.15.1
Flask==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
legacy==0.1.7
MarkupSafe==2.1.5
numpy==1.26.4
openweathermap-requests==0.0.5
pandas==2.2.1
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-cache==1.2.0
six==1.16.0
sqlparse==0.4.4
